HDX-6548 copying core base.html to hdx_theme to change link_for params
removed "controller='organization" and "action='read'" params as it
confuses the url_for logic. Actually it's an HDX controller
(mentioned both in after_map and before_map so that the build_nav_icon
helper function realizes that the current tab is active), not the
core 'organization' controller that is serving this url.