HDX-6990 There were too many dataset IDs in the 'fq' query
switched the solr query to filter by organization instead of the list of datasets ids coming from mixpanel.
The total download number for the org will remain the same.

However, a dataset that was moved from one org to another has no chance of still appearing in the top downloads list.
( This was possible before )