[#8668] Add auth function for recently_changed_packages_activity_list
Fixes #8668

The `recently_changed_packages_activity_list` originally had an auth
function back in CKAN 2.9 when activities were in core. When they were
migrated to a core extension in 2.10 the auth function was lost but it
didn't break the action because it didn't check access. In #7885
(private activities) a `check_access` call was added but not the auth
function. This made the action fail.
This wasn't raised in the tests because all use `call_action` that uses
`ignore_auth: True` by default and so the auth function is never
reached.
This change adds the auth function with the original value (all
allowed). Note that the actual datasets shown will be filtered by
permission labels.

(cherry picked from commit 434334ac07dceb2a04a0231f1c19d0d9fd22d492)